# docs/intel接口文档-中文版本-openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Host Service",
        "description": "主机管理和WebSocket实时通信服务",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/host/hosts/vnc/report": {
            "post": {
                "tags": [
                    "浏览器插件-VNC连接管理"
                ],
                "summary": "上报 VNC 连接结果",
                "description": "处理浏览器插件上报的 VNC 连接结果，更新主机状态并管理执行日志",
                "operationId": "report_vnc_connection_api_v1_host_hosts_vnc_report_post",
                "parameters": [
                    {
                        "name": "accept-language",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Accept-Language 请求头",
                            "title": "Accept-Language"
                        },
                        "description": "Accept-Language 请求头"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VNCConnectionReport",
                                "description": "VNC 连接结果上报数据"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "上报成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Result_VNCConnectionResponse_"
                                },
                                "example": {
                                    "code": 200,
                                    "message": "VNC连接结果上报成功",
                                    "data": {
                                        "host_id": "123",
                                        "connection_status": "success",
                                        "connection_time": "2025/10/15 10:00:00"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "主机不存在或请求数据无效",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 400,
                                    "message": "主机不存在: 123",
                                    "error_code": "HOST_NOT_FOUND"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/host/hosts/vnc/connect": {
            "post": {
                "tags": [
                    "浏览器插件-VNC连接管理"
                ],
                "summary": "获取 VNC 连接信息",
                "description": "获取指定主机的 VNC 连接参数，用于建立 VNC 连接",
                "operationId": "get_vnc_connection_api_v1_host_hosts_vnc_connect_post",
                "parameters": [
                    {
                        "name": "accept-language",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Accept-Language 请求头",
                            "title": "Accept-Language"
                        },
                        "description": "Accept-Language 请求头"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GetVNCConnectionRequest",
                                "description": "获取 VNC 连接信息请求数据"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Result_VNCConnectionInfo_"
                                },
                                "example": {
                                    "code": 200,
                                    "message": "操作成功",
                                    "data": {
                                        "ip": "192.168.101.118",
                                        "port": "5900",
                                        "username": "neusoft",
                                        "***REMOVED***word": "***REMOVED***"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "请求数据无效或 VNC 信息不完整",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 400,
                                    "message": "主机ID格式无效",
                                    "error_code": "INVALID_HOST_ID"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "主机不存在或未启用",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 53001,
                                    "message": "主机不存在或未启用",
                                    "error_code": "HOST_NOT_FOUND"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/host/agent/vnc/report": {
            "post": {
                "tags": [
                    "Agent-硬件信息上报"
                ],
                "summary": "Agent 上报 VNC 连接状态",
                "description": "Agent 上报 VNC 连接状态，系统会根据状态更新主机状态。\n\n    ## 功能说明\n    1. 从 JWT token 中提取 host_id\n    2. 根据 vnc_state 和当前 host_state 更新主机状态\n\n    ## 认证要求\n    - 需要在 Authorization 头中提供有效的 JWT token\n    - Token 格式：`Bearer <token>`\n    - Token 中的 id（从 user_id 或 sub 字段提取）将作为 host_id 使用\n\n    ## 请求参数\n    - `vnc_state`: VNC连接状态（必需）\n        - `1`: 连接成功\n        - `2`: 连接断开\n\n    ## 业务逻辑\n    1. 从 token 中解析 host_id（通过依赖注入自动完成）\n    2. 查询 host_rec 表，验证主机是否存在\n    3. 根据 vnc_state 和当前 host_state 更新状态：\n        - 当 `vnc_state = 1`（连接成功）时：\n            - 如果 `host_state = 1`（已锁定），则修改为 `host_state = 2`（已占用）\n            - 如果 `host_state` 不等于 1，将返回 `VNC_STATE_MISMATCH` 错误\n        - 当 `vnc_state = 2`（连接断开/失败）时：\n            - 不需要做状态判断，直接修改为 `host_state = 0`（空闲）\n            - 同时逻辑删除 `host_exec_log` 表中对应的 host 的有效数据（`del_flag = 1`）\n\n    ## 返回数据\n    - `host_id`: 主机ID\n    - `host_state`: 更新后的主机状态（0=空闲，1=已锁定，2=已占用）\n    - `vnc_state`: 上报的VNC连接状态（1=连接成功，2=连接断开）\n    - `updated`: 是否成功更新\n\n    ## 错误码\n    - `HOST_NOT_FOUND`: 主机不存在或已删除（404，错误码：53001）\n    - `VNC_STATE_MISMATCH`: VNC连接成功但主机状态不匹配（400，错误码：53016）\n        - 当 `vnc_state = 1`（连接成功）时，要求 `host_state = 1`（已锁定）\n        - 如果 `host_state` 不等于 1，将返回此错误\n    - `VNC_CONNECTION_REPORT_FAILED`: 上报处理失败（500，错误码：53020）",
                "operationId": "report_vnc_connection_state_api_v1_host_agent_vnc_report_post",
                "parameters": [
                    {
                        "name": "accept-language",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Accept-Language 请求头",
                            "title": "Accept-Language"
                        },
                        "description": "Accept-Language 请求头"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AgentVNCConnectionReportRequest",
                                "description": "Agent VNC 连接状态上报请求"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "上报成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Result_AgentVNCConnectionReportResponse_"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "请求参数错误或业务逻辑错误",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "validation_error": {
                                        "summary": "请求参数验证失败",
                                        "value": {
                                            "code": 400,
                                            "message": "请求参数验证失败",
                                            "error_code": "VALIDATION_ERROR",
                                            "details": {
                                                "errors": [
                                                    {
                                                        "loc": [
                                                            "body",
                                                            "vnc_state"
                                                        ],
                                                        "msg": "ensure this value is greater than or equal to 1",
                                                        "type": "value_error.number.not_ge"
                                                    }
                                                ]
                                            }
                                        }
                                    },
                                    "vnc_state_mismatch": {
                                        "summary": "VNC连接成功但主机状态不匹配",
                                        "value": {
                                            "code": 53016,
                                            "message": "VNC连接成功，但主机状态不匹配。当前状态：0，需要状态：1（已锁定）",
                                            "error_code": "VNC_STATE_MISMATCH",
                                            "http_status_code": 400,
                                            "details": {
                                                "host_id": 123,
                                                "vnc_state": 1,
                                                "current_host_state": 0,
                                                "required_host_state": 1
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "主机不存在或已删除",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 53001,
                                    "message": "主机不存在: 123",
                                    "error_code": "HOST_NOT_FOUND"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 53020,
                                    "message": "Agent VNC 连接状态上报处理失败",
                                    "error_code": "VNC_CONNECTION_REPORT_FAILED"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AgentVNCConnectionReportRequest": {
                "properties": {
                    "vnc_state": {
                        "type": "integer",
                        "maximum": 2,
                        "minimum": 1,
                        "title": "Vnc State",
                        "description": "VNC连接状态（1=连接成功，2=连接断开）"
                    }
                },
                "type": "object",
                "required": [
                    "vnc_state"
                ],
                "title": "AgentVNCConnectionReportRequest",
                "description": "Agent VNC 连接状态上报请求\n\nvnc_state 说明：\n- 1: 连接成功\n- 2: 连接断开",
                "example": {
                    "vnc_state": 1
                }
            },
...
            "GetVNCConnectionRequest": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id",
                        "description": "主机ID (host_rec.id)"
                    }
                },
                "type": "object",
                "required": [
                    "id"
                ],
                "title": "GetVNCConnectionRequest",
                "description": "获取 VNC 连接信息请求模式"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "Result_AgentVNCConnectionReportResponse_": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "title": "Code",
                        "description": "响应码",
                        "default": 200
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "响应消息",
                        "default": "操作成功"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AgentVNCConnectionReportResponse",
                        "description": "响应数据"
                    },
                    "timestamp": {
                        "type": "string",
                        "title": "Timestamp",
                        "description": "响应时间戳（ISO 8601 格式）"
                    },
                    "locale": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Locale",
                        "description": "语言代码（用于多语言支持）"
                    }
                },
                "type": "object",
                "required": [
                    "data"
                ],
                "title": "Result[AgentVNCConnectionReportResponse]"
            },
...
            "Result_VNCConnectionInfo_": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "title": "Code",
                        "description": "响应码",
                        "default": 200
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "响应消息",
                        "default": "操作成功"
                    },
                    "data": {
                        "$ref": "#/components/schemas/VNCConnectionInfo",
                        "description": "响应数据"
                    },
                    "timestamp": {
                        "type": "string",
                        "title": "Timestamp",
                        "description": "响应时间戳（ISO 8601 格式）"
                    },
                    "locale": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Locale",
                        "description": "语言代码（用于多语言支持）"
                    }
                },
                "type": "object",
                "required": [
                    "data"
                ],
                "title": "Result[VNCConnectionInfo]"
            },
            "Result_VNCConnectionResponse_": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "title": "Code",
                        "description": "响应码",
                        "default": 200
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "响应消息",
                        "default": "操作成功"
                    },
                    "data": {
                        "$ref": "#/components/schemas/VNCConnectionResponse",
                        "description": "响应数据"
                    },
                    "timestamp": {
                        "type": "string",
                        "title": "Timestamp",
                        "description": "响应时间戳（ISO 8601 格式）"
                    },
                    "locale": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Locale",
                        "description": "语言代码（用于多语言支持）"
                    }
                },
                "type": "object",
                "required": [
                    "data"
                ],
                "title": "Result[VNCConnectionResponse]"
            },
...
            "VNCConnectionReport": {
                "properties": {
                    "user_id": {
                        "type": "string",
                        "title": "User Id",
                        "description": "用户ID"
                    },
                    "tc_id": {
                        "type": "string",
                        "title": "Tc Id",
                        "description": "执行测试ID"
                    },
                    "cycle_name": {
                        "type": "string",
                        "title": "Cycle Name",
                        "description": "周期名称"
                    },
                    "user_name": {
                        "type": "string",
                        "title": "User Name",
                        "description": "用户名称"
                    },
                    "host_id": {
                        "type": "string",
                        "title": "Host Id",
                        "description": "主机ID"
                    },
                    "connection_status": {
                        "type": "string",
                        "pattern": "^(success|failed)$",
                        "title": "Connection Status",
                        "description": "连接状态 (success/failed)"
                    },
                    "connection_time": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Connection Time",
                        "description": "连接时间（支持格式：yyyy/MM/dd HH:mm:ss 或 ISO 8601）"
                    }
                },
                "type": "object",
                "required": [
                    "user_id",
                    "tc_id",
                    "cycle_name",
                    "user_name",
                    "host_id",
                    "connection_status",
                    "connection_time"
                ],
                "title": "VNCConnectionReport",
                "description": "VNC 连接结果上报 - 浏览器插件上报VNC连接结果"
            },
            "VNCConnectionResponse": {
                "properties": {
                    "host_id": {
                        "type": "string",
                        "title": "Host Id",
                        "description": "主机ID"
                    },
                    "connection_status": {
                        "type": "string",
                        "title": "Connection Status",
                        "description": "连接状态"
                    },
                    "connection_time": {
                        "type": "string",
                        "title": "Connection Time",
                        "description": "连接时间"
                    }
                },
                "type": "object",
                "required": [
                    "host_id",
                    "connection_status",
                    "connection_time"
                ],
                "title": "VNCConnectionResponse",
                "description": "VNC 连接结果上报响应"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}